Attempt button text background solution
In some instances button text background would retain :hover background-color. This problem would start with a click. This attempted fix uses ::selection to keep button background the same in the case there was a selection (highlighted text).

This problem is hard to replicate so it may be some time before it is known if this fix provides a successful solution.